# openapi/v1.yaml
openapi: 3.1.0
info:
  title: registerwatch
  description: Gambling regulators' public registers, one schema per register, with history.
  contact:
    name: registerwatch
    url: https://github.com/mishmish004/RegisterWatch
  license:
    name: All rights reserved
    url: https://choosealicense.com/no-permission/
  version: 0.2.0
servers:
- url: /
  description: The deployment serving this document
paths:
  /v1/jurisdictions:
    get:
      tags:
      - jurisdictions
      summary: List jurisdictions
      operationId: listJurisdictions
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionPage'
          headers:
            Cache-Control: &id001
              $ref: '#/components/headers/Cache-Control'
            ETag: &id002
              $ref: '#/components/headers/ETag'
            RateLimit-Policy: &id003
              $ref: '#/components/headers/RateLimit-Policy'
            RateLimit: &id004
              $ref: '#/components/headers/RateLimit'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security:
      - ReadToken: []
      - IngestToken: []
  /v1/jurisdictions/{code}:
    get:
      tags:
      - jurisdictions
      summary: Get a jurisdiction with its registers and their freshness
      operationId: getJurisdiction
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
          examples:
          - ch
          title: Code
        description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurisdictionDetail'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/jurisdictions/{code}/changes:
    get:
      tags:
      - changes
      summary: Rows added and removed across a jurisdiction's registers, oldest first
      description: Registers interleave by when their snapshots were recorded. Each register's first complete
        snapshot is its baseline, not a change.
      operationId: listJurisdictionChanges
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
          examples:
          - ch
          title: Code
        description: 'Jurisdiction code, any case, `-` or `_`: `gb`, `US-NJ`, `ca_on`'
      - name: since
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Only changes recorded at or after this RFC 3339 time, with an offset (`2026-10-01T00:00:00Z`)
          title: Since
        description: Only changes recorded at or after this RFC 3339 time, with an offset (`2026-10-01T00:00:00Z`)
      - name: until
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          description: Only changes recorded before this time; after `since`
          title: Until
        description: Only changes recorded before this time; after `since`
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Page size
          default: 100
          title: Limit
        description: Page size
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: '`next_cursor` from the previous page'
          title: Cursor
        description: '`next_cursor` from the previous page'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePage'
          headers:
            Cache-Control: *id001
            ETag: *id002
            RateLimit-Policy: *id003
            RateLimit: *id004
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '504':
          $ref: '#/components/responses/QueryTimeout'
...
  /jurisdictions:
    get:
      tags:
      - read
      summary: List Jurisdictions
      operationId: list_jurisdictions_jurisdictions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response List Jurisdictions Jurisdictions Get
        '400':
          description: A parameter or the path carried a NUL byte
      security:
      - ReadToken: []
      - IngestToken: []
  /jurisdictions/{code}:
    get:
      tags:
      - read
      summary: Get Jurisdiction
      operationId: get_jurisdiction_jurisdictions__code__get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Jurisdiction Jurisdictions  Code  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /jurisdictions/{code}/changes:
    get:
      tags:
      - read
      summary: Get Changes
      operationId: get_changes_jurisdictions__code__changes_get
      security:
      - ReadToken: []
      - IngestToken: []
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
          title: Code
      - name: since
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: ISO date or timestamp
          title: Since
        description: ISO date or timestamp
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 9223372036854775807
          minimum: 0
          default: 200
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Changes Jurisdictions  Code  Changes Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ChangeEvent:
      properties:
        change:
          type: string
          enum:
          - added
          - removed
          title: Change
          description: A changed row is one `removed` and one `added`
        at:
          type: string
          format: date-time
          title: At
          description: When the snapshot that made the change was recorded
        snapshot_id:
          type: integer
          title: Snapshot Id
        register:
          type: string
          title: Register
          description: Register slug, also its Postgres schema
        table:
          type: string
          title: Table
        row:
          $ref: '#/components/schemas/Row'
      type: object
      required:
      - change
      - at
      - snapshot_id
      - register
      - table
      - row
      title: ChangeEvent
      examples:
      - &id005
        change: removed
        at: '2026-10-07T06:00:44Z'
        snapshot_id: 431
        register: gb_ukgc
        table: licences
        row:
          id: 1187
          first_seen_snapshot_id: 412
          first_seen_at: '2026-10-05T06:01:12Z'
          last_seen_at: '2026-10-07T06:00:41Z'
          values:
            account_number: '102'
            licence_number: 000102-N-000000-000
            status: Active
            type: Non-Remote
            activity: Bingo
            start_date: null
            end_date: null
    ChangePage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/ChangeEvent'
          type: array
          title: Data
          description: 'Oldest first: by snapshot, then table, row id, change'
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: ChangePage
      examples:
      - data:
        - *id005
        pagination:
          next_cursor: eyJwIjp7ImsiOls0MzEsImdiX3VrZ2MiLCJsaWNlbmNlcyIsMTE4NywicmVtb3ZlZCJdfSwicyI6IjEyIn0
          has_more: true
          limit: 1
          total: null
    Column:
      properties:
        name:
          type: string
          title: Name
        type:
          type: string
          enum:
          - text
          - date
          - timestamptz
          - int
          - bigint
          - numeric
          - bool
          - text[]
          title: Type
        required:
          type: boolean
          title: Required
          description: Filled in on (nearly) every row of a good snapshot
      type: object
      required:
      - name
      - type
      - required
      title: Column
      examples:
      - name: licence_number
        type: text
        required: true
...
    Freshness:
      properties:
        last_attempt_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Attempt At
          description: Newest ingest attempt, good or not
        last_good_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Good At
          description: Newest complete snapshot; null if there has never been one
        last_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Reason
          description: Why the newest attempt was incomplete, e.g. `PAGE_GAP:3/4 licences:HTTP_503`; null
            if it was complete
        failed_7d:
          type: integer
          title: Failed 7D
          description: Incomplete attempts in the last 7 days
      type: object
      required:
      - last_attempt_at
      - last_good_at
      - last_reason
      - failed_7d
      title: Freshness
      examples:
      - last_attempt_at: '2026-10-07T06:00:41Z'
        last_good_at: '2026-10-07T06:00:41Z'
        last_reason: null
        failed_7d: 0
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Jurisdiction:
      properties:
        code:
          type: string
          title: Code
          description: ISO 3166-1 alpha-2, or a subdivision such as US-NJ
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        registers:
          items:
            $ref: '#/components/schemas/RegisterSummary'
          type: array
          title: Registers
      type: object
      required:
      - code
      - name
      - url
      - registers
      title: Jurisdiction
      examples:
      - &id009
        code: CH
        name: Switzerland
        url: /v1/jurisdictions/ch
        registers:
        - slug: ch_esbk
          jurisdiction: CH
          regulator: Eidgenössische Spielbankenkommission (ESBK)
          name: Blocklist of unauthorised online casino games
          kind: blocklist
          url: /v1/registers/ch_esbk
        - slug: ch_gespa
          jurisdiction: CH
          regulator: Gespa — Interkantonale Geldspielaufsicht
          name: Blocklist of unauthorised lottery and betting sites
          kind: blocklist
          url: /v1/registers/ch_gespa
    JurisdictionDetail:
      properties:
        code:
          type: string
          title: Code
        name:
          type: string
          title: Name
        url:
          type: string
          title: Url
        registers:
          items:
            $ref: '#/components/schemas/RegisterDetail'
          type: array
          title: Registers
      type: object
      required:
      - code
      - name
      - url
      - registers
      title: JurisdictionDetail
      examples:
      - code: CH
        name: Switzerland
        url: /v1/jurisdictions/ch
        registers:
        - slug: ch_esbk
          jurisdiction: CH
          regulator: Eidgenössische Spielbankenkommission (ESBK)
          name: Blocklist of unauthorised online casino games
          kind: blocklist
          homepage: https://www.esbk.admin.ch/en/unauthorised-online-games
          url: /v1/registers/ch_esbk
          tables:
          - &id010
            register: ch_esbk
            name: blocked_domains
            description: null
            columns:
            - name: domain
              type: text
              required: true
            - name: listed_on
              type: date
              required: false
            url: /v1/registers/ch_esbk/tables/blocked_domains
            rows_url: /v1/registers/ch_esbk/tables/blocked_domains/rows
          freshness: null
    JurisdictionPage:
      properties:
        data:
          items:
            $ref: '#/components/schemas/Jurisdiction'
          type: array
          title: Data
        pagination:
          $ref: '#/components/schemas/Pagination'
      type: object
      required:
      - data
      - pagination
      title: JurisdictionPage
      examples:
      - data:
        - *id009
        pagination:
          next_cursor: null
          has_more: false
          limit: 20
          total: 20
    Pagination:
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: Pass as `cursor` for the next page; null on the last page
        has_more:
          type: boolean
          title: Has More
        limit:
          type: integer
          title: Limit
          description: Page size this response was read with
        total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Total
          description: Matching items; only with `include_total=true` on endpoints that offer it, else
            null
      type: object
      required:
      - next_cursor
      - has_more
      - limit
      title: Pagination
      examples:
      - next_cursor: eyJwIjp7Im8iOjEwMH0sInMiOiI0ZjFkIn0
        has_more: true
        limit: 100
        total: null
...
    RegisterDetail:
      properties:
        slug:
          type: string
          title: Slug
        jurisdiction:
          type: string
          title: Jurisdiction
        regulator:
          type: string
          title: Regulator
        name:
          type: string
          title: Name
        kind:
          type: string
          enum:
          - licensees
          - blocklist
          title: Kind
        homepage:
          type: string
          title: Homepage
        url:
          type: string
          title: Url
        tables:
          items:
            $ref: '#/components/schemas/TableSchema'
          type: array
          title: Tables
        freshness:
          anyOf:
          - $ref: '#/components/schemas/Freshness'
          - type: 'null'
          description: Null when the database cannot be reached
      type: object
      required:
      - slug
      - jurisdiction
      - regulator
      - name
      - kind
      - homepage
      - url
      - tables
      - freshness
      title: RegisterDetail
      examples:
      - slug: ch_esbk
        jurisdiction: CH
        regulator: Eidgenössische Spielbankenkommission (ESBK)
        name: Blocklist of unauthorised online casino games
        kind: blocklist
        homepage: https://www.esbk.admin.ch/en/unauthorised-online-games
        url: /v1/registers/ch_esbk
        tables:
        - *id010
        freshness:
          last_attempt_at: '2026-10-07T06:00:41Z'
          last_good_at: '2026-10-07T06:00:41Z'
          last_reason: null
          failed_7d: 0
...
    RegisterSummary:
      properties:
        slug:
          type: string
          title: Slug
        jurisdiction:
          type: string
          title: Jurisdiction
        regulator:
          type: string
          title: Regulator
        name:
          type: string
          title: Name
        kind:
          type: string
          enum:
          - licensees
          - blocklist
          title: Kind
        url:
          type: string
          title: Url
      type: object
      required:
      - slug
      - jurisdiction
      - regulator
      - name
      - kind
      - url
      title: RegisterSummary
      examples:
      - slug: gb_ukgc
        jurisdiction: GB
        regulator: Gambling Commission
        name: Public register of gambling businesses
        kind: licensees
        url: /v1/registers/gb_ukgc
    Row:
      properties:
        id:
          type: integer
          title: Id
          description: Stable for as long as the row is unchanged; a changed row is a new row (and the
            old one is removed). Pages are keyed on it
        first_seen_snapshot_id:
          type: integer
          title: First Seen Snapshot Id
          description: The complete snapshot that first contained this row
        first_seen_at:
          type: string
          format: date-time
          title: First Seen At
          description: When the first complete snapshot containing this row was recorded
        last_seen_at:
          type: string
          format: date-time
          title: Last Seen At
          description: When the newest complete snapshot still containing it was recorded
        values:
          additionalProperties:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - type: string
              format: date-time
            - type: string
              format: date
            - items:
                type: string
              type: array
            - type: 'null'
          type: object
          title: Values
          description: The register's own columns, as published; keys are the table's declared columns
      type: object
      required:
      - id
      - first_seen_snapshot_id
      - first_seen_at
      - last_seen_at
      - values
      title: Row
      examples:
      - &id013
        id: 1187
        first_seen_snapshot_id: 412
        first_seen_at: '2026-10-05T06:01:12Z'
        last_seen_at: '2026-10-07T06:00:41Z'
        values: &id012
          account_number: '102'
          licence_number: 000102-N-000000-000
          status: Pending
          type: Non-Remote
          activity: Bingo
          start_date: null
          end_date: null
...
    TableSchema:
      properties:
        register:
          type: string
          title: Register
          description: Register slug, also its Postgres schema
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        columns:
          items:
            $ref: '#/components/schemas/Column'
          type: array
          title: Columns
        url:
          type: string
          title: Url
        rows_url:
          type: string
          title: Rows Url
      type: object
      required:
      - register
      - name
      - description
      - columns
      - url
      - rows_url
      title: TableSchema
      examples:
      - register: gb_ukgc
        name: licences
        description: One row per licence and activity; licence_number's last segment is a version
        columns:
        - name: account_number
          type: text
          required: true
        - name: licence_number
          type: text
          required: true
        - name: status
          type: text
          required: true
        - name: start_date
          type: date
          required: false
        url: /v1/registers/gb_ukgc/tables/licences
        rows_url: /v1/registers/gb_ukgc/tables/licences/rows
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    FieldError:
      examples:
      - field: limit
        location: query
        message: Input should be less than or equal to 1000
      properties:
        field:
          title: Field
          type: string
        location:
          description: '`query`, `path` or `header`'
          title: Location
          type: string
        message:
          title: Message
          type: string
      required:
      - field
      - location
      - message
      title: FieldError
      type: object
    Problem:
      additionalProperties: true
      description: RFC 9457 Problem Details. Extension members may be added per type.
      examples:
      - detail: 'limit: Input should be less than or equal to 1000'
        errors:
        - field: limit
          location: query
          message: Input should be less than or equal to 1000
        instance: /v1/registers/gb_ukgc/tables/licences/rows
        request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
        status: 400
        title: Invalid parameter
        type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
      properties:
        type:
          description: Stable identifier of the problem type; its docs say what to do
          format: uri
          title: Type
          type: string
        title:
          title: Title
          type: string
        status:
          title: Status
          type: integer
        detail:
          description: What went wrong with this request, for a human
          title: Detail
          type: string
        instance:
          description: The request path
          format: uri-reference
          title: Instance
          type: string
        request_id:
          description: Also in the `X-Request-Id` response header and the server's logs
          title: Request Id
          type: string
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          default: null
          description: Each invalid parameter, for 400s
          title: Errors
      required:
      - type
      - title
      - status
      - detail
      - instance
      - request_id
      title: Problem
      type: object
  securitySchemes:
    ReadToken:
      type: http
      description: '`READ_TOKEN`. Needed on read routes only when the server sets one; when it does not,
        reads are open. The ingest token reads too.'
      scheme: bearer
    IngestToken:
      type: http
      description: '`INGEST_TOKEN`. Starts and lists ingest runs, and reads. A read token here is a 403.'
      scheme: bearer
  responses:
    BadRequest:
      description: 'A parameter is invalid; `errors` names each one. Types: `invalid-parameter`, `unknown-filter-column`,
        `invalid-cursor`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#invalid-parameter
            title: Invalid parameter
            status: 400
            detail: 'limit: Input should be less than or equal to 1000'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
            errors:
            - field: limit
              location: query
              message: Input should be less than or equal to 1000
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    Unauthenticated:
      description: 'The bearer token is missing, or is not one this route takes. Types: `unauthenticated`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#unauthenticated
            title: Authentication required
            status: 401
            detail: the bearer token is not valid here
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        WWW-Authenticate:
          description: '`Bearer realm="registerwatch"`, with `error="invalid_token"` when a token was
            sent but is not valid (RFC 6750)'
          schema:
            type: string
        RateLimit-Policy: *id003
        RateLimit: *id004
...
    NotFound:
      description: 'An identifier in the path or body does not exist. Types: `jurisdiction-not-found`,
        `register-not-found`, `table-not-found`, `row-not-found`, `ingest-run-not-found`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#jurisdiction-not-found
            title: Unknown jurisdiction
            status: 404
            detail: 'unknown jurisdiction ''zz''; known: au, be, ca, ...'
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
...
    TooManyRequests:
      description: 'Rate limit exceeded; wait `Retry-After` seconds. Types: `rate-limited`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#rate-limited
            title: Too many requests
            status: 429
            detail: 60 requests per minute for /v1/search; retry in 12 s
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    InternalError:
      description: 'An unexpected error; quote `request_id` when reporting it. Types: `internal`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#internal
            title: Internal error
            status: 500
            detail: something went wrong on our side; quote request_id if you report it
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        RateLimit-Policy: *id003
        RateLimit: *id004
    ServiceUnavailable:
      description: 'The database cannot be reached (retry after `Retry-After` seconds), or ingest is not
        configured on this deployment. Types: `database-unavailable`, `ingest-disabled`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#database-unavailable
            title: Database unavailable
            status: 503
            detail: the database cannot be reached right now; retry shortly
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    QueryTimeout:
      description: 'A query ran past the statement timeout and was cancelled; narrow the request, or retry
        after `Retry-After` seconds. Types: `query-timeout`'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          example:
            type: https://github.com/mishmish004/RegisterWatch/blob/main/docs/problems.md#query-timeout
            title: Query took too long
            status: 504
            detail: the query was cancelled for taking too long; narrow it and retry
            instance: /v1/registers/gb_ukgc/tables/licences/rows
            request_id: 0199c1a8-7c3e-7a52-9d1e-5b6f0c2a4e11
      headers:
        Retry-After: *id016
        RateLimit-Policy: *id003
        RateLimit: *id004
    NotModified:
      description: Unchanged since the `ETag` sent in `If-None-Match`; no body
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
        Cache-Control:
          $ref: '#/components/headers/Cache-Control'
        RateLimit-Policy: *id003
        RateLimit: *id004
  headers:
    ETag:
      description: This representation's version (weak). Send it back in `If-None-Match` and the answer
        is a 304 with no body while the data is unchanged. New data shows within 30 seconds of its snapshot.
      schema:
        type: string
      example: W/"9f2c5d0e..."
    Cache-Control:
      description: '`public, max-age=300` when reads are open, `private, max-age=300` when the deployment
        requires a read token; `no-store` on ingest operations'
      schema:
        type: string
    RateLimit-Policy:
      description: 'The quota this request counts against (draft-ietf-httpapi-ratelimit-headers): `"read";q=600;w=60`
        is 600 requests per 60 seconds. Per bearer token, or per client address without one. Defaults:
        `read` 600, `search` (search and domains) 60, `ingest` (starting runs) 10 a minute.'
      schema:
        type: string
      example: '"search";q=60;w=60'
    RateLimit:
      description: 'Where the client stands: `r` requests left, more available in `t` seconds'
      schema:
        type: string
      example: '"search";r=41;t=1'